+++ playwright-tests/pdfmake-examples/styling_properties.test.ts
@@ -45,6 +45,24 @@ const document: TDocumentDefinitions = {
       ],
     },
     { text: ['Hello', { text: '1, 2', sub: true }, ' World'] },
+    {
+      text: 'Customize word break:',
+      pageBreak: 'before',
+    },
+    {
+      text: 'DefaultLine\n"BreakBehaviour" "ForATextWithVeryVery" "LongLongWords"',
+      fontSize: 30,
+    },
+    {
+      text: '\n\n',
+      fontSize: 30,
+    },
+    {
+      text: 'BreakAll\n"LineBreakBehaviour" "ForATextWithVeryVery" "LongLongWords"',
+      fontSize: 30,
+      // @ts-expect-error need to add to @types/pdfmake
+      wordBreak: 'break-all',
+    },
   ],
 }